Preparation for tokenizer testing
